pyproject.toml
[project]
name = "agentdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.3",
    "langgraph>=1.0.5",
    "prompt-toolkit>=3.0.52",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]
